Remove code from header files, simplify compilation
* Pull the code out of USBstreamUtils.h into a new file USBstreamUtils.cxx
so that we aren't having crazy #include of the implementation itself.

* As a consequence, there is no longer a global gaibu_debug_msg variable
that everything can write into willy nilly with or without actually
logging.  Found a couple of places where it was written into, but then
not used, and replaced these with actual logging commands.

* Excised Gaibu itself, and replaced the implemetation with writing to
stdout.  It also still logs to syslog.  It also attempts to send mail
for bad errors, but that doesn't work since it refers to Double Chooz
files.

* USBstream.cxx was being compiled into a shared library as well as a
normal object file.  I don't see why, so I removed that.  If there's
a reason that still applies, it could easily be reinstated.