Simplify OVhit structure.
Don't define a typedef and a class that wraps it and nothing else.
Then don't pretend that the data members are private when we
expose them raw through a call to write().  And don't use a
pointer to an instance of this struct when a plain struct will do.

I am assuming that sizeof(pair<char, short>) == sizeof( a struct
made of a char and a short).  On my system this is true.

Also, what on Earth happens when you write() a std::pair<char, short>)?
It depends how that is stored, doesn't it?  Is that defined?  I tested
write()ing a std::pair<char, short> and a struct{ char a; short b}.
Both have size 4 (on my system!) and are padded the same way, but the
pair came through with junk in the padding bits.  That's probably
not going to be consistent.  Anyway, I think my change has the
same result, but both versions are worrisome.  If we have three
bytes of data, I'd rather write out three, not four.