Deobfuscate another typedef.
Change DataVectorIt to DataVector::iterator so that it is plain what
it is.  Remove unnecessary header gaurds (it's legal to typedef multiple
times).  Remove comment that just spells out what the type of the
variable is now that it is clear from the name.

Now that there's only one typedef left, I kinda want to get rid of it
too, but it is the only one that actually helps readability by being
much shorter than what it is replacing, so we'll see.