Pull out next segment of database interaction.
Now read the number of non-fan-in USBs in the same function as the
number of total USBs, and switch between them with a bool.

Also note that we never want the number of fan-in USBs, but rather
the number of non-fan-ins, so use *that* as the global variable.

Note disturbing cast of the number of non-fan-in USBs to void* to
initialize a TThread.  What does *that* mean?