Pull out next part of DB interaction.
And discover that it issues warnings about how it's going to change
things, but doesn't change them (trigger mode and threshold, specifically).
I've left this be, but added XXX comments to it.  Apparently this functionality
wasn't actually used?  Maybe we should just abort in this case, or maybe
there should be no such command line options.  Or maybe we should *always*
get this from the command line and not the database?