Load up hard-coded data to avoid DB for now.
Use the mysqldump I got from Camillo to populate hard-coded data
that I can use for testing.  It is currently not clear to me that
this data needs to be read from a database or if a single config
file would do the trick for ProtoDUNE.  So this will probably
all be changed again.