Document and simplify GetDir().
Simplify GetDir() and give its arguments meaningful names.  They are
now bools and the function returns a bool.  Don't try to closedir()
a directory that didn't get opened. use while() instead of do{if(){}}while().
Split baffling if statement apart until I could understand it and
change the names.  Don't set a local variable that's about to go out
of scope to NULL.