Reverted USBstream.cxx to original.
I had finally gotten the code to *run*, but was getting
the error that a module number was 1536.  Since this is
6 << 8, I suspected that one of the many changes I made
to USBstream.cxx introduced a decoding bug.  Since
'module' was correctly set in check_data(), the problem
is somewhere downstream of that.  But rather than hunt it
down, just dump the old version in, modulo the interface
changes that I made.

Now it runs up to complaining that a directory that
certainly exists doesn't exist.

Later I will need to more carefully update USBstream.cxx.