Improve errors. Allow setting output dir.
* Use strerror() so the user (or, actually, the programmer
at this point) can know why things are failing.

* Change the meaning of the command line switch -e so that
instead of setting just the number in a hardcoded path, it
sets the whole path.

* Set fname in main() to something non-empty to begin with.
It looks like this is supposed to be set before it is used,
but that isn't happening for me.  This needs to be fixed, but
at least now we aren't trying to open a file named "".