Restore compatibility to the old (problematic) format.
Pad the output file with magic words where there used to be
vtable pointers written out.  This keeps everthing aligned as
it was in the Double Chooz version and makes it visible where
the junk is when you view a hexdump.  Next I'm going to redesign
the file format to not have these at all.