Made BuildEvent() far more readable. std::.
* Changed the names of the major data structures in BuildEvent
so that it is possible to understand what is going on.  For
one thing, the input data no longer has a name containing "Out".
It's a vector of packets, so make that clear.

* Restructured the loop so that the two things being looped over
are manifestly locked together.

* Pulled out the stuff that operates on the first packet from the
loop and have it do that before the loop instead of being a
special case of the loop.

The deal with what's called "OutIndex", and which gets the module
number assigned correctly is still murky to me.  The logic which
creates this vector is quite tangled.  I think that ultimately
with only one USB stream in protoDUNE (?) it is doing a lot of
work to add zero to a number.

* Abolished "using namespace std".  That can lead to teeth gnashing
when you find you have named a variable something you didn't know
was in std::, and there are a *lot* of things in std::.

* Along with that, abolished use of std::cout.  Either we send
text to log_msg() or we don't need to print it.  Generally made
program less chatty about internal details.  (No more "creating thread!").