Drop support for database interaction. NB!
Stop pretending that we're going to interact with a database
and save 500 lines of code.  I'm pretty sure that I want to
read in the basic statistics of the detector using a text
config file and do logging via text stream.  Another layer
outside this program can pick that up as desired with much less
code.

Obviously, if dropping this was a bad idea, you can reverse
cherry-pick this commit.  But I'm pretty sure that even if I
decided to use a database for this in the future, it would not
be mysql.  Probably something way easier to use like sqlite.