Drop concepts of run mode and run type.
* Drop run mode.  This removes support for reprocessing and recovery modes.
It's not clear to me that these are needed nor that they work. The code
complexity incurred is huge, by which I mean I can't understand what the
code relating to them is trying to do.  I suspect that if these capabilities
are needed, they can be handled in a script that calls this program, given
that the input and output file names can be set more directly, which is
on my TODO list.

* Drop run type.  All runs are now physics.  The *only* consequence of this
is that they have the letter "P" in the output file name.  I plan to make
the output file name directly settable with logic about how it is set moved
into a shell script as needed.