Simplify file handling.
* Set permissions to 644 and 755, not 666 and 777.

* Move GetDir into EventBuilder.cxx and remove unused option.

* Drop concept of DAQ disk number.  We will use directory names.

* Don't avoid a few directory listings at the cost having a global
variable which is mutated by multiple functions.  Just list the
directory when we need to.

* Improve some print statements.

* Remove checks that can't fail about whether the file names that
we just sorted into place to match up with the USB numbers are
matched up with the USB numbers.

* Change option -H to mean the input directory.

* Do not create the unused output directory /processed.

* Reduce the number of references to RunNumber.  I'm not sure how
run number will be handled and it may be removed.

* Put processed files in a subdirectory of the input directory
called /decoded rather than a sibling directory.  Now we don't
need to make assumptions about the path of the input directory.