Refactoring for readability.
Pull huge hunks of code out of main() willy nilly in an attempt
to understand the logic of the main loop.  I think I'm getting
close.  The main loop at least now fits on one screen and can
be examined.

I may well have broken some of the logic that tries to decide how
or when to close output files.  But it was impossible to
understand that logic as is, and I have no idea if I want it to
do what it was doing.