Apply offsets to ADC hits.
Previously the code applied timing offsets to latch and trigger
box hits.  Then since we aren't going to use those for
protoDUNE, I removed that, although some remnants remained.
But now apply timing offsets to ADC hits, which, as per
Camillo, is the right thing to do.

I have restored the old (difficult to understand) logic for
applying offsets to the low and high words of the time stamp
in USBStream.cxx.  However, I have used a much simplified
method for getting the offsets stored in the first place.
No more pointer arithmetic.