Read/decode/synch loop now transparent.
It is now possible to read the loop that reads in files,
decodes them, then moves the data into the arrays from which
they will be merged together.  I'm not sure it's doing the
right thing, though.  It produces identical output for my
test data set as the code did before, but I'm not at all
confident that my test data is sufficient to test this.  See
new XXX comment in the code.