Tighten up restart procedure in decode().
I can see no reason why we should sleep between passes over the
file when we get half of a Unix timestamp packet or whatever it
is that is happening in check_debug() that triggers this.  The
sleeping would make sense if this were some sort of a read error,
but it isn't.

Also, instead of making it the responsibility of decode() in
EventBuilder.cxx to restart the decoding, short circuit this
by jumping back to the top of USBStream::decode().