Document first stage of decoding, no flush.
* I have reverse-engineered the input file format, something
I didn't know I needed to do until today, because I thought
the input file format was what is documented in Matt Toups'
thesis.  But it's not, instead the 24 bit words documented
therein are spread over 32 bits in the file with two-bit
counters at the front of each octet.

* Now that I'm starting to understand what's going on, I have
removed flush_extra, which didn't do much as it was.